Fixed type inconsistency.
SDL_GetTicks() returns Uint32, but I was using the long int type,
which could mean signed or unsigned, not to mention any size. Changed
missionCompleteTimer and timeTaken to Uint32 for consistency.

+++ src/engine.h
@@ -64,11 +64,11 @@ typedef struct Engine_ {
 	int targetIndex;
 
 	// Mission completion timer (allows for 4 seconds before leaving sector)
-	long missionCompleteTimer;
+	Uint32 missionCompleteTimer;
 
 	// Times the mission normally
 	Uint32 counter2;
-	long int timeTaken; // In seconds
+	Uint32 timeTaken; // In seconds
 
 	// For missions with a time limit
 	int timeMission;
